requirements.txt
flask==2.0.1
flask-cors==3.0.10
tensorflow==2.7.0
opencv-python==4.5.3.56
numpy==1.19.5
pillow==8.3.2